// node/configure/type/jr-cyber-station.d.ts
/* tslint:disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

/**
 * 自然言語による、人間が見て分かりやすい名前を設定する。通知メールの件名などで使用される。
 */
export type NoName = string;
export type URL = string;
export type URL1 = string;
export type NoName1 = number;
export type NoName3 = string;
export type NoName4 = string;
export type NoName5 = string;
export type NoName6 = string;
export type NoName2 = {
  depature: NoName3;
  arrival: NoName4;
  date: NoName5;
  time?: NoName6;
}[];

export interface JR {
  title: NoName;
  station_url: URL;
  search_url: URL1;
  search_interval: NoName1;
  search: NoName2;
}